requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
fn7-sdk>=1.0.0
firebase-admin>=6.0.0
pydantic>=2.0.0

